+++ CHANGELOG.rst
@@ -5,6 +5,12 @@ LATEST CHANGES
 - Added Docker files (`docker` folder) ready to create Docker images for desktop host devices: Ubuntu 22.04, ROS2 Foxy, ZED SDK v3.8.2.
 - Changed the default depth mode to `PERFORMANCE` (1) for a faster node start in the case that the NEURAL depth AI model is not present.
 
+2022-12-19
+----------
+- Change `general.grab_resolution` and `general.pub_resolution` from integer to string.
+- Add new `LOW` value for `general.pub_resolution` (half the `MEDIUM` output resolution).
+- Fix distortion model (see Issue [#128](https://github.com/stereolabs/zed-ros2-wrapper/issues/128))
+
 2022-12-12
 ----------
 - Added diagnostic warning if the frequency of the camera grabbing thread is minor than the selected `general.grab_frame_rate` value.